refactor: extract capability::diagnostics (probe-failure classification, TLS rendering)
Second step of the capability.rs module split, stacked on the
state.rs extraction. Pure lift-and-shift: ConnFailure, explicit/record/
reset_explicit_probe_failure, error_chain, find_rustls_cause,
describe_rustls_error, and cert_trust_hint (plus their tests) move into
capability/diagnostics.rs unchanged. capability.rs re-exports the
external-facing pieces and privately imports the rest, since its
still-inline probe code calls into them.

Depends only on the state.rs split (not on any sibling extracted in
this step or later), so it lands second.
